untitled/src/main/java/com/epam/tree/TreeExecutor.java
package com.epam.tree;

import java.util.ArrayList;
import java.util.List;
import java.util.Objects;

public class TreeExecutor {

    public List<Integer> traverse(Tree tree) {
        List<Integer> result = new ArrayList<>();

        Tree leftTree = tree.getLeft();
        Tree rightTree = tree.getRight();

        if (Objects.nonNull(leftTree)) {
            result.addAll(traverse(leftTree));
        }
        if (Objects.nonNull(rightTree)) {
            result.addAll(traverse(rightTree));
        }

        result.add(tree.getValue());

        return result;
    }

}
